WIP experiment instance entry packing; not working

+++ src/mesh/InstanceSSBO.h
@@ -17,8 +17,8 @@ namespace mesh {
 
         GLuint u_entityIndex;
         GLuint u_meshIndex;
-        GLuint u_materialIndex;
-        GLuint u_shapeIndex{ 0};
+        uint16_t u_materialIndex;
+        uint16_t u_shapeIndex{ 0 };
     };
 #pragma pack(pop)
 }
